Settopbox/requirements.txt
apturl==0.5.2
bcrypt==3.1.7
bidict==0.22.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
call==0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.1
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.10.2
google-auth==2.14.1
google-cloud==0.34.0
google-cloud-speech==2.16.2
googleapis-common-protos==1.56.4
greenlet==2.0.0.post0
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
keyboard==0.13.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
monotonic==1.5
MouseInfo==0.1.3
netifaces==0.10.6
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.6.0.66
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
playsound==1.3.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycairo==1.16.2
pycoral==2.0.0
pycparser==2.21
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.6
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
pyserial==3.5
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-engineio==4.3.4
python-socketio==5.7.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
SpeechRecognition==3.8.1
svgwrite==1.4.3
systemd-python==234
tflite-runtime==2.5.0.post1
typing-extensions==4.4.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==1.4.2
Werkzeug==2.2.2
xkit==0.0.0
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.1
